Revert "drm/i915: Mark the FB's bo as busy when we pageflip"
This reverts commit d8446056008f14ac2cb36cc531a6756a380e49e9.

We are trying to nail down the cause for a recent series of test
crashes in Chrome. Since this is basically the only graphics
change in that time frame, let's try to revert this for a couple
of days and see if our tests heal.

BUG=chromium-os:37791
TEST=compiles, works on Alex.

Change-Id: I5a1086389be56ca2603371b4f89d7a1e3d2d74ee
Reviewed-on: https://gerrit.chromium.org/gerrit/41263
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>