BACKPORT: fs: ecryptfs: initialize payload_len in keystore.c
This is meant to remove a compiler warning.  It should not make any
functional change.

payload_len should be initialized when it is passed to
write_tag_64_packet() as a pointer.  If that call fails, this function
should return early, and payload_len won't be used.

The warning is:
fs/ecryptfs/keystore.c:1168:28: warning: 'payload_len' may be used
uninitialized in this function [-Wmaybe-uninitialized]

BUG=chromium-os:5542
TEST=emerge-x86-generic chromeos-kernel, this warning should not appear.

Change-Id: Ia4fb282b3b32eadecc7d2756c6a701864480618b
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41580
Reviewed-by: Mandeep Singh Baines <msb@chromium.org>