BACKPORT: ALSA: hda - Add input jack mode enum controls to generic parser
Just like the jack mode enum ctls for output jacks, add the support
for similar enum ctls for input pins to control the bias Vref.
The new controls will be added when spec->add_in_jack_modes is set
either by the codec driver or by a hint string.

Note that ground and 100% vrefs are excluded from the list for
simplicity, currently.  We may add a new flag to allow them, too.
But I guess it's easier to put a value override in the pinfix in such
a case.

Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit 29476558deb3017993366a3b0b45dff2acad495b)

BUG=chrome-os-partner:19100
TEST=Verified the working of HDMI audio with both stereo and headset by building
     the ChromeOS image along with entire 3.9 HDMI driver patches and four
     additional patches from drm/i915.