exynos-gsc: remove check for gsc_m2m_active during resume
There is a check during gsc resume which will return if the device
was not active before suspend. This check is not correct since we
do wait for the device to become idle before suspending. So the gsc
device will always be idle during resume and we end up not resuming
the gsc device. This patch removes this incorrect check.

Signed-off-by: Prathyush K <prathyush.k@samsung.com>