UPSTREAM: of: fix incorrect return value of of_find_matching_node_and_match()
The of_find_matching_node_and_match() function incorrectly sets the matched
entry to 'matches' when the compatible value of a node matches one of the
possible values. This results in incorrectly selecting the the first entry in
the 'matches' list as the matched entry. Fix this by noting down the result of
the call to of_match_node() and setting that as the matched entry.

BUG=chromium:266008
TEST=Tested on spring

Change-Id: I85dd71ce16319fbec8128a2b77e232674c69e4d8
Signed-off-by: Thomas Abraham <thomas.abraham@linaro.org>
Signed-off-by: Rob Herring <rob.herring@calxeda.com>
Signed-off-by: Sean Paul <seanpaul@chromium.org>
(cherry picked from commit dc71bcf1b99c265fffdb1e8c60e7933612f3bca7)
Reviewed-on: https://chromium-review.googlesource.com/177756