BACKPORT: ALSA: hda - Synchronize the power state at the end of codec init
Put the power state synchronization at the end of the parsing of
codec.  This is necessary when the power filter is changed during the
codec probe.  Since the first power-up sequence is performed without
the special filter, all widgets are supposed to be ON at this point.

Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit b9c590bbf1d7621c3f9feb6ac0992d638244d0b1)

BUG=chrome-os-partner:19100
TEST=Verified the working of HDMI audio with both stereo and headset by building
     the ChromeOS image along with entire 3.9 HDMI driver patches and four
     additional patches from drm/i915.