UPSTREAM: HID: multitouch: use the callback "report" instead of sequential events
Nexio 42" devices requires to rely on the HID field Contact
Count to compute the valid values. However, this field is
most of the time at the end of the report, meaning that we
need to get the all report parsed before processing it.

This patch does not introduce functional changes.

BUG=chromium:318205
TEST=None

Change-Id: I79ee788e5e492270720d81586d2f62b75d37ac8e
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
Reviewed-on: https://chromium-review.googlesource.com/183262
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Yufeng Shen <miletus@chromium.org>
Commit-Queue: Yufeng Shen <miletus@chromium.org>