UPSTREAM: cfg80211: fix BSS list hidden SSID lookup
Upstream commit 4593c4cbe1c96b3995727dc42f6aa103f4ff5afc.

When trying to find a hidden SSID, the lookup function
is done wrong; the code is trying to combine the two
lookups into one, and as a consequence doesn't always
find the entry at all. To understand this, consider a
case where multiple BSS entries with the same channel
and BSSID exist but have different SSID length. Then
comparing against the probe response SSID length is
bound to cause problems since the hidden one might be
either zeroed out or zero-length.

To fix this we need to do two lookups for the two ways
to hide SSIDs.

BUG=chrome-os-partner:20691
TEST=tested that connectivity works over ath9k and iwl7000

Change-Id: I98551535909248c5b31f0e1ed13f0068de05506b
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/59760
Reviewed-by: Paul Stewart <pstew@chromium.org>