CHROMIUM: rtc: Allow opening rtc by OF node
This new function allows a device tree phandle to be used to locate an rtc
class device. The name (rtc0, rtc1) depends on the device init order, which
means it cannot be used within the device tree.

BUG=chrome-os-partner:10617
TEST=build and boot to kernel prompt on snow
Using charger-manager, select either rtc0 or rtc1 as the rtc to use for
the timer alarms, and see that charger-manager wakes up and behaves
correctly in suspend_again() in each case.

Change-Id: Ia6206b26cd74297f16b8e68caf97b6f390f4a7c5
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43013
Reviewed-by: Benson Leung <bleung@chromium.org>