CHROMIUM: Update tps65090 device tree entry for charger
Add a sub-node for charger. We do not specify a name for the regulator
since that would enable this function of the driver, although it would
not really have any effect. Still, we don't want this until/unless
charger-manager is enabled.

BUG=chrome-os-partner:10617
TEST=build and boot to kernel prompt on snow
With other changes, see that the charger can be enabled and disabled under
charger-manager control.

Change-Id: Ida12beedae31e848e775b7083ea2c998a35bda32
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43024