exynos: i2c: Correct suspend/resume GPIO request/free
The gpio_request() is done in the resume handler, but the gpio_free()
is done in the suspend_noirq() handler. These do not match.

Where the system comes up partially from suspend and then goes back to
sleep (via the suspend_again handler), the suspend_noirq() handler
is called but the resume handler is not. This means that we free
the GPIOs but do not request them again, and thus we end up freeing
them again on each suspend_again.

Correct this by putting the gpio_free() into the suspend() handler
instead of suspend_noirq(). The effect is the same, for a full
suspend/resume cycle

BUG=chrome-os-partner:10617
TEST=build and boot to kernel prompt on snow
Suspend and resume and see that the i2c bus resumes correctly.
Using charger-manager, take the device through several suspend_again
cycles and see that warnings about multiple gpio_free() calls are
no longer present in the logs.
Change-Id: If11f02675d948959d81fbc5cf05cbb38605b4791
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43015
Reviewed-by: Doug Anderson <dianders@chromium.org>