CHROMIUM: charger-manager: Move temperature thresholds to pdata
Rather than require a function to handle temperature thresholds,
specify the thresholds so that charger-manager can do the right thing
automatically. This is is easier for clients to use, and more
device-tree-friendly

BUG=chrome-os-partner:10617
TEST=build and boot to kernel prompt on snow
See that charger-manager works as before. There is currently no test
for temperature thresholds, so I manually printed them out after a
later change to see that they are correct.
Change-Id: Ide0efff77c3c4e7bd468abf9df3b80884014e28d
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43016
Reviewed-by: Benson Leung <bleung@chromium.org>