CHROMIUM: exynos: properly manage "regulator" for HSIC reset
We were violating something in the documentation for regulator_put,
namely:
  Note: drivers must ensure that all regulator_enable calls made on
  this regulator source are balanced by regulator_disable calls prior
  to calling this function.

We now keep the regulator around until we've disabled it.  That keeps
the regulator in the right state.

At the same time, add some error checking to regulator calls.

BUG=chromium-os:38403
TEST=on Snow with 3G modem, do a suspend/resume cycle and check in
"lsusb" that the modem (VID/PID 1410:a021) and the camera (VID/PID
2232:1037) are still there after resume.
On Spring, check we can still boot from USB.
TEST=Check state of hsichub-reset-l at boot and after resume:
1. cd /sys/kernel/debug/regulator/hsichub-reset-l
2. grep "" *
3. Check open_count of 1 and use_count of 1.

Change-Id: I31012c15e30c43a1259f15ffd3592ee68cb0b772
Signed-off-by: Doug Anderson <dianders@chromium.org>
Signed-off-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/42621
Reviewed-by: Todd Broch <tbroch@chromium.org>